ELIS-8437: Fix broken upgrade step in elis_program.

+++ elis/program/db/upgrade.php
@@ -605,16 +605,18 @@ function xmldb_elis_program_upgrade($oldversion=0) {
         foreach ($customfields as $id => $unused) {
             $field = new field($id);
             $field->load();
-            $manual = new field_owner($field->owners['manual']);
-            $control = $manual->param_control;
-            $options = $manual->param_options;
-            if (!empty($options) && empty($manual->param_options_source) && ($control == 'menu' || $control == 'checkbox')) {
-                $options = str_replace("\r", '', $options); // strip CRs
-                $options = preg_replace("/\n+/", "\n", $options);
-                $manual->param_options = rtrim($options, "\n");
-                $manual->save();
-                // Remove any empty defaults
-                $DB->delete_records_select($field->data_table(), "contextid IS NULL AND fieldid = ? AND data = ''", array($id));
+            if (isset($field->owners['manual'])) {
+                $manual = new field_owner($field->owners['manual']);
+                $control = $manual->param_control;
+                $options = $manual->param_options;
+                if (!empty($options) && empty($manual->param_options_source) && ($control == 'menu' || $control == 'checkbox')) {
+                    $options = str_replace("\r", '', $options); // strip CRs
+                    $options = preg_replace("/\n+/", "\n", $options);
+                    $manual->param_options = rtrim($options, "\n");
+                    $manual->save();
+                    // Remove any empty defaults
+                    $DB->delete_records_select($field->data_table(), "contextid IS NULL AND fieldid = ? AND data = ''", array($id));
+                }
             }
         }
         upgrade_plugin_savepoint($result, 2012062902, 'elis', 'program');
